refactor: move list logic to application layer
