+++ src/application/repositories/PokemonRepository.ts
@@ -1,14 +1,6 @@
 import { Effect } from 'effect';
 import type { Pokemon } from '@/domain/pokemon';
 
-export type ListQuery = {
-  q?: string;
-  legendary?: boolean;
-  page?: number;
-  pageSize?: number;
-  sort?: string;
-};
-
 export class NotFound extends Error {
   readonly _tag = 'NotFound';
   constructor(message: string) {
@@ -17,12 +9,7 @@ export class NotFound extends Error {
 }
 
 export interface PokemonRepository {
-  list(query: ListQuery): Effect.Effect<{
-    total: number;
-    page: number;
-    pageSize: number;
-    data: Pokemon[];
-  }, Error>;
+  getAll(): Effect.Effect<ReadonlyArray<Pokemon>, Error>;
   getById(id: number): Effect.Effect<Pokemon, Error>;
   getByIdWithSimilar(
     id: number,
