+++ tests/integration/repo-list.test.ts
@@ -2,19 +2,28 @@
 import { describe, it, expect } from 'vitest';
 import { Effect, Either } from 'effect';
 import { PokemonRepositoryCsv } from '@/infrastructure/repositories/PokemonRepositoryCsv';
+import { list } from '@/application/pokemon/list';
 
 const FIXTURE = 'data/pokemon_fixture_30.csv';
 
-describe('整合：Repository 查詢/排序/分頁', () => {
+describe('整合：Application 查詢/排序/分頁', () => {
   it('legendary=true 過濾 + 多欄排序', async () => {
     const repo = new PokemonRepositoryCsv(FIXTURE);
-    const eff = repo.list({ legendary: true, page: 1, pageSize: 100, sort: 'bst:desc,name:asc' });
-    const r = await Effect.runPromise(Effect.either(eff));
+    const eff = list(repo, {
+      legendary: 'true',
+      page: '1',
+      pageSize: '100',
+      sort: 'bst:desc,name:asc',
+    });
+    const r = await Effect.runPromise(eff);
     expect(Either.isRight(r)).toBe(true);
     if (Either.isRight(r)) {
       const out = r.right;
+      expect(out.page).toBe(1);
+      expect(out.pageSize).toBe(100);
+      expect(out.total).toBeGreaterThan(0);
       expect(out.data.length).toBeGreaterThan(0);
-      expect(out.data.every(p => p.legendary)).toBe(true);
+      expect(out.data.every((p) => p.legendary)).toBe(true);
     }
   });
 });
